+++ RiotSwiftUI/Modules/Room/Composer/ViewModel/ComposerViewModelProtocol.swift
@@ -23,4 +23,6 @@ protocol ComposerViewModelProtocol {
     var textFormattingEnabled: Bool { get set }
     var eventSenderDisplayName: String? { get set }
     var placeholder: String? { get set }
+    
+    func dismissKeyboard()
 }