+++ RiotSwiftUI/Modules/UserSessions/UserSessionOverview/Test/Unit/UserSessionOverviewViewModelTests.swift
@@ -22,7 +22,7 @@ import XCTest
 class UserSessionOverviewViewModelTests: XCTestCase {
     func test_whenVerifyCurrentSessionProcessed_completionWithVerifyCurrentSessionCalled() {
         let sessionInfo = createUserSessionInfo()
-        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: MockUserSessionOverviewService())
+        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: MockUserSessionOverviewService(), settingsService: MockUserSessionSettings())
         
         XCTAssertEqual(sut.state.isPusherEnabled, nil)
         var modelResult: UserSessionOverviewViewModelResult?
@@ -35,7 +35,7 @@ class UserSessionOverviewViewModelTests: XCTestCase {
     
     func test_whenViewSessionDetailsProcessed_completionWithShowSessionDetailsCalled() {
         let sessionInfo = createUserSessionInfo()
-        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: MockUserSessionOverviewService())
+        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: MockUserSessionOverviewService(), settingsService: MockUserSessionSettings())
 
         XCTAssertEqual(sut.state.isPusherEnabled, nil)
         var modelResult: UserSessionOverviewViewModelResult?
@@ -49,7 +49,7 @@ class UserSessionOverviewViewModelTests: XCTestCase {
     func test_whenViewSessionDetailsProcessed_toggleAvailablePusher() {
         let sessionInfo = createUserSessionInfo()
         let service = MockUserSessionOverviewService(pusherEnabled: true)
-        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service)
+        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service, settingsService: MockUserSessionSettings())
 
         XCTAssertTrue(sut.state.remotelyTogglingPushersAvailable)
         XCTAssertEqual(sut.state.isPusherEnabled, true)
@@ -62,7 +62,7 @@ class UserSessionOverviewViewModelTests: XCTestCase {
     func test_whenViewSessionDetailsProcessed_toggleNoPusher() {
         let sessionInfo = createUserSessionInfo()
         let service = MockUserSessionOverviewService(pusherEnabled: nil)
-        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service)
+        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service, settingsService: MockUserSessionSettings())
 
         XCTAssertTrue(sut.state.remotelyTogglingPushersAvailable)
         XCTAssertEqual(sut.state.isPusherEnabled, nil)
@@ -75,7 +75,7 @@ class UserSessionOverviewViewModelTests: XCTestCase {
     func test_whenViewSessionDetailsProcessed_remotelyTogglingPushersNotAvailable() {
         let sessionInfo = createUserSessionInfo()
         let service = MockUserSessionOverviewService(pusherEnabled: true, remotelyTogglingPushersAvailable: false)
-        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service)
+        let sut = UserSessionOverviewViewModel(sessionInfo: sessionInfo, service: service, settingsService: MockUserSessionSettings())
 
         XCTAssertFalse(sut.state.remotelyTogglingPushersAvailable)
         XCTAssertEqual(sut.state.isPusherEnabled, true)