+++ RiotTests/UserSessionsDataProviderTests.swift
@@ -113,7 +113,7 @@ private class MockSession: MXSession {
 /// A mock `MXCrypto` that can override the `canCrossSign` state.
 private class MockCrypto: MXLegacyCrypto {
     let canCrossSign: Bool
-    override var crossSigning: MXCrossSigning! { MockCrossSigning(canCrossSign: canCrossSign) }
+    override var crossSigning: MXCrossSigning { MockCrossSigning(canCrossSign: canCrossSign) }
     
     init(canCrossSign: Bool) {
         self.canCrossSign = canCrossSign